+++ src/main/java/com/codetaylor/mc/artisantools/common/event/ConstructModEventHandler.java
@@ -6,6 +6,7 @@ import com.codetaylor.mc.artisantools.common.material.JsonConfigFileGenerator;
 import com.codetaylor.mc.artisantools.common.material.JsonConfigFileReader;
 import com.codetaylor.mc.artisantools.common.pack.GeneratedPackRemover;
 import com.codetaylor.mc.artisantools.common.pack.PackGenerator;
+import com.codetaylor.mc.artisantools.common.recipe.DataRecipeTemplateMap;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
 
@@ -14,6 +15,8 @@ public class ConstructModEventHandler {
   private final JsonConfigFileGenerator<DataCustomMaterialList> materialFileGenerator;
   private final JsonConfigFileReader<DataCustomMaterialList> materialFileReader;
   private final CustomMaterialListPopulator customMaterialListPopulator;
+  private final JsonConfigFileGenerator<DataRecipeTemplateMap> recipeFileGenerator;
+  private final JsonConfigFileReader<DataRecipeTemplateMap> recipeFileReader;
   private final GeneratedPackRemover generatedPackRemover;
   private final PackGenerator resourcePackGenerator;
   private final PackGenerator dataPackGenerator;
@@ -22,6 +25,8 @@ public class ConstructModEventHandler {
       JsonConfigFileGenerator<DataCustomMaterialList> materialFileGenerator,
       JsonConfigFileReader<DataCustomMaterialList> materialFileReader,
       CustomMaterialListPopulator customMaterialListPopulator,
+      JsonConfigFileGenerator<DataRecipeTemplateMap> recipeFileGenerator,
+      JsonConfigFileReader<DataRecipeTemplateMap> recipeFileReader,
       GeneratedPackRemover generatedPackRemover,
       PackGenerator resourcePackGenerator,
       PackGenerator dataPackGenerator
@@ -30,6 +35,8 @@ public class ConstructModEventHandler {
     this.materialFileGenerator = materialFileGenerator;
     this.materialFileReader = materialFileReader;
     this.customMaterialListPopulator = customMaterialListPopulator;
+    this.recipeFileGenerator = recipeFileGenerator;
+    this.recipeFileReader = recipeFileReader;
     this.generatedPackRemover = generatedPackRemover;
     this.resourcePackGenerator = resourcePackGenerator;
     this.dataPackGenerator = dataPackGenerator;
@@ -41,6 +48,8 @@ public class ConstructModEventHandler {
     this.materialFileGenerator.generate();
     this.materialFileReader.read();
     this.customMaterialListPopulator.populate();
+    this.recipeFileGenerator.generate();
+    this.recipeFileReader.read();
     this.generatedPackRemover.remove();
     this.resourcePackGenerator.generate();
     this.dataPackGenerator.generate();