+++ src/main/java/com/codetaylor/mc/artisantools/common/recipe/DataRecipeTemplateMap.java
@@ -1,0 +1,8 @@
+package com.codetaylor.mc.artisantools.common.recipe;
+
+import java.util.Map;
+
+public class DataRecipeTemplateMap {
+
+  public Map<String, DataRecipeTemplate> recipes;
+}