+++ src/main/java/com/codetaylor/mc/artisantools/common/recipe/RecipeValidationException.java
@@ -1,0 +1,10 @@
+package com.codetaylor.mc.artisantools.common.recipe;
+
+public class RecipeValidationException
+    extends Exception {
+
+  public RecipeValidationException(String message) {
+
+    super(message);
+  }
+}