+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # 2.1.0
 
 **Added:**
+  * Config json file for modifying recipe templates [(#1)](https://github.com/codetaylor/artisan-tools-1.16/issues/1)
   * Paintbrush tool [(#2)](https://github.com/codetaylor/artisan-tools-1.16/issues/2)
 
 ---